Refactor attribute descriptor handling and update imports for consistency across modules
